LayoutTests/fast/overflow/height-during-simplified-layout.html

<html>
<body>

<p>This test passes if you see a green box below.</p>

<div id="scrollable" style="height: 0; overflow-y: auto; padding-bottom: 200px; background-color: green">
  <div style="position: relative; height: 400px; background-color: red">
    <div id="node-to-hide" style="position: absolute;">hello</div>
  </div>
</div>
<script src="../js/resources/js-test-pre.js"></script>
<script>
document.body.offsetLeft;
document.getElementById("node-to-hide").style.display = "none";
document.getElementById("scrollable").scrollTop = "400";
shouldBe("document.getElementById('scrollable').scrollTop", "400");
shouldBe("document.getElementById('scrollable').scrollHeight", "600");
</script>
</body>
</html>
