LayoutTests/fast/overflow/height-during-simplified-layout-expected.txt

PASS document.getElementById('scrollable').scrollTop is 400
PASS document.getElementById('scrollable').scrollHeight is 600
This test passes if you see a green box below.
